sync review comments
