+++ at_client/test/functional_test/atclient_dedicated_connection_test.dart
@@ -11,8 +11,6 @@ void main() {
     var preference = getAlicePreference(atsign);
     await AtClientImpl.createClient(atsign, 'me', preference);
     AtClientImpl atClient = await AtClientImpl.getClient(atsign);
-    atClient.getSyncManager().init(atsign, preference,
-        atClient.getRemoteSecondary(), atClient.getLocalSecondary());
     await atClient.getSyncManager().sync();
     // To setup encryption keys
     await setEncryptionKeys(atsign, preference);
