+++ at_client/test/functional_test/set_encryption_keys.dart
@@ -8,8 +8,6 @@ Future<void> setEncryptionKeys(
   try {
     await AtClientImpl.createClient(atsign, 'me', atClientPreference);
     var atClient = await AtClientImpl.getClient(atsign);
-    atClient.getSyncManager().init(atsign, atClientPreference,
-        atClient.getRemoteSecondary(), atClient.getLocalSecondary());
     var metadata = Metadata();
     metadata.namespaceAware = false;
     var result;
