+++ at_client/test/samples/at_client_check_sync.dart
@@ -7,12 +7,9 @@ import 'test_util.dart';
 void main() async {
   try {
     var atSign = '@alice🛠';
-    var preference = TestUtil.getAlicePreference();
     await AtClientImpl.createClient(
         atSign, 'me', TestUtil.getAlicePreference());
     var atClient = await AtClientImpl.getClient(atSign);
-    atClient.getSyncManager().init(atSign, preference,
-        atClient.getRemoteSecondary(), atClient.getLocalSecondary());
     var result = await atClient.getSyncManager().isInSync();
     print(result);
   } on Exception catch (e, trace) {
